// src/types/globalTypes.ts
// ​​​academicProgramName: "Civil Engineering (4 Years, Bachelor of Technology)"
// ​​​closeRank: 12911
// ​​​collegeId: "edc6b2e8-c08a-477c-9d2a-de2c3d35ccd9"
// ​​​createdAt: "2025-02-26T08:35:49.691Z"
// ​​​exam: "ADVANCED"
// ​​​gender: "Gender-Neutral"
// ​​​id: "56f41a87-05a9-46ed-ab6c-94447e1e0c1d"
// ​​​institute: "Indian Institute of Technology Bhubaneswar"
// ​​​openRank: 9462
// ​​​quota: "AI"
// ​​​round: 1
// ​​​seatType: "OPEN"
// ​​​type: "JOSSA"
// ​updatedAt: "2025-02-26T08:35:49.691Z"
// ​​year: 2023

export type Orcr = {
    id: string
    institute: string
    academicProgramName: string
    quota: string
    seatType: string
    gender: string
    openRank: number
    closeRank: number
    year: number
    round: number
    type: string
    exam: string
    collegeId: string
    createdAt: string
    updatedAt: string
}